Add `useCreateWorkflow` hook
